Settings and Watchdog Changes
Updated the settings share so that it no longer uses an enumerable for values. Acts as a dictionary now (like a settings collection should). Updated some watchdog configuration routines and added new settings to toggle it on or off. Changed around some view models which used the settings configurations as well.